+++ FlashCap/FrameProcessors/IgnoreDroppingProcessor.cs
@@ -30,7 +30,7 @@ namespace FlashCap.FrameProcessors
 
         public override sealed void OnFrameArrived(
             CaptureDevice captureDevice,
-            IntPtr pData, int size, long timestampMicroseconds)
+            IntPtr pData, int size, double timestampMicroseconds)
         {
             if (Interlocked.Increment(ref isin) == 1)
             {
